% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellQCMetrics.R
\name{addPerCellQCMetrics}
\alias{addPerCellQCMetrics}
\title{Add QC metrics}
\usage{
addPerCellQCMetrics(object, ...)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[=perCellQCMetrics]{perCellQCMetrics}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Add QC metrics
}
\seealso{
\itemize{
\item \code{\link{perCellQCMetrics}}
\item \code{\link[scrapper:rna_quality_control]{computeRnaQcMetrics}}
}
}
