% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGraph.R
\name{clusterGraph}
\alias{clusterGraph}
\alias{clusterGraph.igraph}
\title{Graph-based clustering}
\usage{
clusterGraph(object, ...)

\method{clusterGraph}{igraph}(
  object,
  method = NULL,
  ...,
  resolution = 1L,
  objective = NULL,
  steps = 4L,
  seed = NULL
)
}
\arguments{
\item{object}{A \code{\link[igraph:make_graph]{graph}} object.}

\item{...}{Additional arguments passed on to
\code{\link[igraph:cluster_louvain]{cluster_louvain}},
\code{\link[igraph:cluster_leiden]{cluster_leiden}} or
\code{\link[igraph:cluster_walktrap]{cluster_walktrap}}.}

\item{method}{String specifying the community detection method to use.
Options are multi-level (\code{"multilevel"}/\code{"louvain"}), Walktrap (\code{"walktrap"})
or Leiden (\code{"leiden"}) (default).}

\item{resolution}{Numeric scalar specifying the resolution to use for
\code{multi-level} or \code{Leiden} clustering.}

\item{objective}{String specifying the objective function to use for Leiden
clustering: "cpm" or "modularity" (default).}

\item{steps}{Integer scalar specifying the number of steps to use for
\code{Walktrap} clustering.}

\item{seed}{Integer scalar specifying the seed to use for \code{multi-level} or
\code{Leiden} clustering.}
}
\value{
A factor of the cluster assignments. Each method may also return
additional attributes.
\itemize{
\item For method=\code{"multilevel"}/\code{"louvain"}, we have:
\itemize{
\item \code{levels}, a list of integer vectors with cluster assignments for each
cell at each level. Assignments are sorted by decreasing resolution (i.e.,
fewer, larger clusters).
\item \code{modularity}, a numeric vector containing the modularity of each level.
\item \code{best}, the level with the lowest modularity.
}
\item For method=\code{"leiden"}, we have:
\itemize{
\item \code{quality}: a numeric scalar containing the quality of the clustering
(either the modularity or a related score).
}
\item For method=\code{"walktrap"}, we have:
\itemize{
\item \code{merges}: an integer matrix specifying how the clusters were merged to
obtain membership. Each row corresponds to a merge step and contains the
IDs of the temporary clusters (not the same as those in membership).
\item \code{modularity}: a numeric vector containing the modularity before and
after each merge step.
}
}
}
\description{
Graph-based clustering
}
