% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffusionMap.R
\name{diffusionMapRoots}
\alias{diffusionMapRoots}
\title{Find roots in a DiffusionMap object}
\usage{
diffusionMapRoots(object, ...)
}
\arguments{
\item{object}{A \code{\link[destiny:DiffusionMap-class]{DiffusionMap}} object.}

\item{start, ends}{The \code{start} and \code{ends} cluster identity. start must have a
length \code{1L}, while the length of ends ranges from \code{1L} to \code{2L}. All \code{start}
and \code{ends} must exist in \code{ref}. If \code{ends} is \code{NULL}, this only check if one
tip is in the start cluster.}

\item{ref}{All cell identity of \code{object}, this must have the same length of
\code{object@d}.}

\item{n_root}{The number of nodes we should test.}
}
\value{
An integer index.
}
\description{
Find roots in a DiffusionMap object. If we know the \code{start} and \code{end} of a
root, we then check if the tips of the root are in the start and ends node we
already known (see \code{\link[destiny:find_tips]{find_tips}}). This function in this
way check \code{n_root} nodes with the top \code{DPT} in the start cluster and return
roots meet this criteria.
}
