% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarySizeFactors.R
\name{librarySizeFactors}
\alias{librarySizeFactors}
\alias{librarySizeFactors.default}
\alias{librarySizeFactors.DelayedArray}
\alias{librarySizeFactors.SummarizedExperiment}
\alias{librarySizeFactors.Seurat}
\title{Compute library size factors}
\usage{
librarySizeFactors(object, ...)

\method{librarySizeFactors}{default}(object, ..., threads = NULL)

\method{librarySizeFactors}{DelayedArray}(object, ...)

\method{librarySizeFactors}{SummarizedExperiment}(object, ..., assay = "counts")

\method{librarySizeFactors}{Seurat}(object, ..., assay = NULL, layer = "counts")
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, a single thread will be used.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}
}
\description{
Compute library size factors
}
