% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNormCounts.R
\name{logNormCounts}
\alias{logNormCounts}
\alias{logNormCounts.default}
\alias{logNormCounts.SummarizedExperiment}
\alias{logNormCounts.Seurat}
\title{Normalize the count matrix}
\usage{
logNormCounts(object, ...)

\method{logNormCounts}{default}(
  object,
  size_factors = NULL,
  ...,
  log = 2L,
  pseudo_count = 1L,
  center_size_factors = TRUE,
  block = NULL,
  mode = NULL
)

\method{logNormCounts}{SummarizedExperiment}(
  object,
  size_factors = NULL,
  ...,
  assay = "counts",
  name = "logcounts"
)

\method{logNormCounts}{Seurat}(object, ..., assay = NULL, layer = "counts", name = "data")
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{size_factors}{A numeric vector of length equal to the number of cells
in \code{object}, containing positive size factors for all cells.}

\item{log}{Numeric scalar specifying the base of the log-transformation. If
\code{NULL}, won't do log-transformation.}

\item{pseudo_count}{Numeric scalar specifying the positive pseudo-count to
add before any log-transformation. Ignored if \code{log = NULL}.}

\item{center_size_factors}{Logical scalar indicating whether size factors
should be centered at unity before being used.}

\item{block}{Vector or factor of length equal to the number of cells,
specifying the batch of origin for each cell. Alternatively \code{NULL} if all
cells belong to the same batch. Ignored if \code{center_size_factors = FALSE}.}

\item{mode}{String specifying how to scale size factors across blocks.
\code{"lowest"} will scale all size factors by the the lowest per-block average.
\code{"per-block"} will center the size factors in each block separately. Ignored
if \code{center_size_factors = FALSE}.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{name}{String specifying the name to be used to store the assay name in
the \code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays}} or the layer name in
\code{\link[SeuratObject:AssayData]{SetAssayData}}.}

\item{layer}{Name of the layer to get from the assay data.}
}
\description{
Apply scaling normalization to a count matrix to obtain log-transformed normalized expression values.
}
\seealso{
\link[scrapper:normalizeCounts]{normalizeCounts}
}
