% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGeneVar.R
\name{modelGeneVar}
\alias{modelGeneVar}
\alias{modelGeneVar.SingleCellExperiment}
\alias{modelGeneVar.Seurat}
\alias{modelGeneVar.default}
\title{Model per-gene variances in expression}
\usage{
modelGeneVar(object, ...)

\method{modelGeneVar}{SingleCellExperiment}(object, ..., assay = "logcounts")

\method{modelGeneVar}{Seurat}(object, ..., assay = NULL, layer = "data")

\method{modelGeneVar}{default}(
  object,
  block = NULL,
  block_weight_policy = NULL,
  ...,
  variable_block_weight = c(0, 1000),
  mean_filter = TRUE,
  min_mean = 0.1,
  transform = TRUE,
  span = 0.3,
  use_min_width = FALSE,
  min_width = 1,
  min_window_count = 200,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for
each cell in \code{x}. Alternatively \code{NULL} if all cells are from the same block.}

\item{block_weight_policy}{A string of \code{"variable"}, \code{"equal"}, or \code{"none"}
specifying the policy to use for weighting different blocks when computing
the average for each statistic Only used if \code{block} is not \code{NULL}.}

\item{variable_block_weight}{Numeric vector of length 2, specifying the
parameters for variable block weighting. The first and second values are used
as the lower and upper bounds, respectively, for the variable weight
calculation. Only used if \code{block} is not \code{NULL} and \code{block_weight_policy = "variable"}.}

\item{mean_filter}{Logical scalar indicating whether to filter on the means
before trend fitting.}

\item{min_mean}{Numeric scalar specifying the minimum mean of genes to use in
trend fitting.  Only used if \code{mean_filter=TRUE}.}

\item{transform}{Logical scalar indicating whether a quarter-root
transformation should be applied before trend fitting.}

\item{span}{Numeric scalar specifying the span of the LOWESS smoother.
Ignored if \code{use_min_width=TRUE}.}

\item{use_min_width}{Logical scalar indicating whether a minimum width
constraint should be applied to the LOWESS smoother.  Useful to avoid
overfitting in high-density intervals.}

\item{min_width}{Minimum width of the window to use when
\code{use_min_width=TRUE}.}

\item{min_window_count}{Minimum number of observations in each window. Only
used if \code{use_min_width=TRUE}.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, a single thread will be used.}
}
\value{
A data frame of class \code{scend_modelGeneVar} with the columns \code{means},
\code{variances}, \code{fitted} \code{residuals}, each of which is a numeric vector
containing the statistic of the same name across all genes.

If \code{block} is supplied, each of the column vectors described above contains
the average across all blocks. The data frame will also contain \code{per_block}
attribute, a list of data frames containing the equivalent statistics for
each block.
}
\description{
Compute the variance in (log-)expression values for each gene,
and model the trend in the variances with respect to the mean.
}
