% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellQCMetrics.R
\name{perCellQCMetrics}
\alias{perCellQCMetrics}
\alias{perCellQCMetrics.default}
\alias{perCellQCMetrics.SingleCellExperiment}
\alias{perCellQCMetrics.Seurat}
\title{Quality control for RNA count data}
\usage{
perCellQCMetrics(object, ...)

\method{perCellQCMetrics}{default}(object, subsets = NULL, threads = NULL, ...)

\method{perCellQCMetrics}{SingleCellExperiment}(object, ..., assay = "counts")

\method{perCellQCMetrics}{Seurat}(object, ..., assay = NULL, layer = "data")
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{subsets}{List of vectors specifying gene subsets of interest, typically for control-like features like mitochondrial genes or spike-in transcripts.
Each vector may be logical (whether to keep each row), integer (row indices) or character (row names).}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, a single thread will be used.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}
}
\value{
A data frame with the columns \code{sum}, \code{detected}, \verb{subsets_*}, each of
which is a numeric vector containing the QC metrics for all cells.
}
\description{
Compute per-cell QC metrics from an initialized matrix of RNA counts,
and use the metrics to suggest filter thresholds to retain high-quality cells.
}
\seealso{
\code{\link[scrapper:rna_quality_control]{computeRnaQcMetrics}}
}
