% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCluster.R
\name{quickCluster}
\alias{quickCluster}
\alias{quickCluster.SingleCellExperiment}
\alias{quickCluster.Seurat}
\title{Quick clustering of cells with \code{scrapper}}
\usage{
quickCluster(object, ...)

\method{quickCluster}{SingleCellExperiment}(
  object,
  size_factors = NULL,
  d = 50L,
  scale = FALSE,
  subset_row = NULL,
  n_hvgs = 2000,
  prop_hvgs = NULL,
  from_residuals = FALSE,
  extra_work = 7,
  iterations = 1000,
  seed = NULL,
  realized = TRUE,
  ...,
  threads = NULL
)

\method{quickCluster}{Seurat}(
  object,
  size_factors = NULL,
  d = 50L,
  scale = FALSE,
  subset_row = NULL,
  n_hvgs = 2000,
  prop_hvgs = NULL,
  from_residuals = FALSE,
  extra_work = 7,
  iterations = 1000,
  seed = NULL,
  realized = TRUE,
  ...,
  threads = NULL
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[=clusterSNNGraph]{clusterSNNGraph}}
  \describe{
    \item{\code{}}{}
  }}

\item{size_factors}{A numeric vector of length equal to the number of cells
in \code{object}, containing positive size factors for all cells.}

\item{d}{Integer scalar specifying the number of top PCs to obtain.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{subset_row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If \code{NULL}, all
features in \code{object} are used.}

\item{n_hvgs}{Integer specifying the number of top genes to retain.}

\item{prop_hvgs}{Numeric scalar specifying the proportion of genes to report
as HVGs.}

\item{from_residuals}{Logical scalar indicating whether to compute the PC
scores from the residuals in the presence of a blocking factor. By default,
the residuals are only used to compute the rotation matrix, and the original
expression values of the cells are projected onto this new space. Only used
if \code{block} is not \code{NULL}.}

\item{extra_work}{Integer scalar specifying the extra dimensions for the
IRLBA workspace.}

\item{iterations}{Integer scalar specifying the maximum nunber of restart iterations for IRLBA.}

\item{seed}{Integer specifying the seed to use for the initial random vector
in \code{IRLBA} and for \code{multi-level} or \code{Leiden} clustering.}

\item{realized}{Logical scalar indicating whether to realize \code{object} into an
optimal memory layout for \code{IRLBA}. This improves computation time at the cost
of increased memory usage.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\code{\link[parallel:detectCores]{detectCores}}.}
}
\description{
Quick clustering of cells with \code{scrapper}
}
\seealso{
\itemize{
\item \code{\link{logNormCounts}}
\item \code{\link{runPCA}}
\item \code{\link{clusterSNNGraph}}
}
}
