% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMNN.R
\name{quickMNN}
\alias{quickMNN}
\alias{quickMNN.SingleCellExperiment}
\alias{quickMNN.Seurat}
\title{Quick fastMNN with \code{scrapper}}
\usage{
quickMNN(object, ...)

\method{quickMNN}{SingleCellExperiment}(
  object,
  block,
  size_factors = NULL,
  ...,
  k = 15L,
  n_mads = 3L,
  robust_iterations = 2,
  robust_trim = 0.25,
  mass_cap = NULL,
  order = NULL,
  reference_policy = NULL,
  BNPARAM = AnnoyParam(),
  threads = NULL,
  name = "corrected"
)

\method{quickMNN}{Seurat}(
  object,
  block,
  size_factors = NULL,
  ...,
  k = 15L,
  n_mads = 3L,
  robust_iterations = 2,
  robust_trim = 0.25,
  mass_cap = NULL,
  order = NULL,
  reference_policy = NULL,
  BNPARAM = AnnoyParam(),
  threads = NULL,
  name = "corrected"
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{
  Arguments passed on to \code{\link[=runPCA]{runPCA}}
  \describe{
    \item{\code{}}{}
  }}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for
each cell in \code{object}.}

\item{size_factors}{A numeric vector of length equal to the number of cells
in \code{x}, containing positive size factors for all cells.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{n_mads}{Numeric scalar specifying the number of median absolute
deviations to use for removing outliers in the center-of-mass calculations.}

\item{robust_iterations}{Integer scalar specifying the number of iterations
for robust calculation of the center of mass.}

\item{robust_trim}{Numeric scalar in [0, 1) specifying the trimming
proportion for robust calculation of the center of mass.}

\item{mass_cap}{Integer scalar specifying the cap on the number of
observations to use for center-of-mass calculations on the reference dataset.
A value of \verb{100,000} may be appropriate for speeding up correction of very
large datasets. If \code{NULL}, no cap is used.}

\item{order}{Vector containing levels of \code{block} in the desired merge
order. If \code{NULL}, a suitable merge order is automatically determined.}

\item{reference_policy}{String specifying the policy to use to choose the
first reference batch. This can be based on the largest batch (\code{"max-size"}),
the most variable batch (\code{"max-variance"}), the batch with the largest
residual sum of squares (\code{"max-rss"}), or the first specified input
(\code{"input"}). Only used for automatic merges, i.e., when \code{order=NULL}.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the nearest-neighbor algorithm to use.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\code{\link[parallel:detectCores]{detectCores}}.}

\item{name}{String specifying the name to be used to store the result in the
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} or
\code{\link[SeuratObject:Seurat-class]{reductions}} of the output.}
}
\description{
Quick fastMNN with \code{scrapper}
}
\seealso{
\itemize{
\item \code{\link{logNormCounts}}
\item \code{\link{runPCA}}
\item \code{\link{runMNN}}
}
}
