% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH5AD.R
\name{readH5AD}
\alias{readH5AD}
\title{Read H5AD}
\usage{
readH5AD(file, delayed = TRUE)
}
\arguments{
\item{file}{A string of \code{h5ad} file path.}

\item{delayed}{A logical value indicating whether assays should be loaded as
\code{\link[HDF5Array:HDF5ArraySeed-class]{HDF5Array}}/
\code{\link[HDF5Array:H5SparseMatrix-class]{H5SparseMatrix}}
from the \code{HDF5Array} package.}
}
\value{
A
\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment} object.
}
\description{
Read H5AD
}
