% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMNN.R
\name{runMNN}
\alias{runMNN}
\alias{runMNN.default}
\alias{runMNN.SingleCellExperiment}
\alias{runMNN.Seurat}
\title{Fast mutual nearest neighbors correction with \code{scrapper}}
\usage{
runMNN(object, ...)

\method{runMNN}{default}(
  object,
  block,
  k = 15L,
  ...,
  n_mads = 3L,
  robust_iterations = 2,
  robust_trim = 0.25,
  mass_cap = NULL,
  order = NULL,
  reference_policy = NULL,
  BNPARAM = AnnoyParam(),
  threads = NULL
)

\method{runMNN}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  name = "corrected"
)

\method{runMNN}{Seurat}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  layer = NULL,
  name = "corrected"
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} methods.
}}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for
each cell in \code{object}.}

\item{k}{Integer scalar specifying the number of neighbors to use when
identifying MNN pairs.}

\item{n_mads}{Numeric scalar specifying the number of median absolute
deviations to use for removing outliers in the center-of-mass calculations.}

\item{robust_iterations}{Integer scalar specifying the number of iterations
for robust calculation of the center of mass.}

\item{robust_trim}{Numeric scalar in [0, 1) specifying the trimming
proportion for robust calculation of the center of mass.}

\item{mass_cap}{Integer scalar specifying the cap on the number of
observations to use for center-of-mass calculations on the reference dataset.
A value of \verb{100,000} may be appropriate for speeding up correction of very
large datasets. If \code{NULL}, no cap is used.}

\item{order}{Vector containing levels of \code{block} in the desired merge
order. If \code{NULL}, a suitable merge order is automatically determined.}

\item{reference_policy}{String specifying the policy to use to choose the
first reference batch. This can be based on the largest batch (\code{"max-size"}),
the most variable batch (\code{"max-variance"}), the batch with the largest
residual sum of squares (\code{"max-rss"}), or the first specified input
(\code{"input"}). Only used for automatic merges, i.e., when \code{order=NULL}.}

\item{BNPARAM}{A \link[BiocNeighbors]{BiocNeighborParam} object specifying the nearest-neighbor algorithm to use.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\code{\link[parallel:detectCores]{detectCores}}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{name}{String specifying the name to be used to store the result in the
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} or
\code{\link[SeuratObject:Seurat-class]{reductions}} of the output.}

\item{layer}{Name of the layer to get from the assay data.}
}
\description{
Fast mutual nearest neighbors correction with \code{scrapper}
}
\seealso{
\code{\link[scrapper:correctMnn]{correctMnn}}
}
