% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\name{runPCA}
\alias{runPCA}
\alias{runPCA.default}
\alias{runPCA.SingleCellExperiment}
\alias{runPCA.Seurat}
\title{Perform Principal components analysis with \code{scrapper}}
\usage{
runPCA(object, ...)

\method{runPCA}{default}(
  object,
  n_dim = 50L,
  scale = FALSE,
  subset_row = NULL,
  ...,
  block = NULL,
  block_weight_policy = NULL,
  variable_block_weight = c(0, 1000),
  from_residuals = FALSE,
  extra_work = 7,
  iterations = 1000,
  seed = NULL,
  realized = TRUE,
  threads = NULL
)

\method{runPCA}{SingleCellExperiment}(
  object,
  ...,
  assay = "logcounts",
  dimred = NULL,
  n_dimred = NULL,
  name = "PCA"
)

\method{runPCA}{Seurat}(
  object,
  ...,
  assay = NULL,
  layer = "data",
  dimred = NULL,
  n_dimred = NULL,
  name = "PCA"
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{n_dim}{Integer scalar specifying the number of top PCs to obtain.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{subset_row}{Integer, logical or character vector specifying which
features to use in the PCA (e.g., highly variable genes). If \code{NULL}, all
features in \code{object} are used.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for
each cell in \code{x}. Alternatively \code{NULL} if all cells are from the same block.}

\item{block_weight_policy}{A string of \code{"variable"}, \code{"equal"}, or \code{"none"}
specifying the policy to use for weighting different blocks when computing
the average for each statistic Only used if \code{block} is not \code{NULL}.}

\item{variable_block_weight}{Numeric vector of length 2, specifying the
parameters for variable block weighting. The first and second values are used
as the lower and upper bounds, respectively, for the variable weight
calculation. Only used if \code{block} is not \code{NULL} and \code{block_weight_policy = "variable"}.}

\item{from_residuals}{Logical scalar indicating whether to compute the PC
scores from the residuals in the presence of a blocking factor. By default,
the residuals are only used to compute the rotation matrix, and the original
expression values of the cells are projected onto this new space. Only used
if \code{block} is not \code{NULL}.}

\item{extra_work}{Integer scalar specifying the extra dimensions for the
IRLBA workspace.}

\item{iterations}{Integer scalar specifying the maximum number of restart iterations for IRLBA.}

\item{seed}{Integer scalar specifying the seed for the initial random vector in IRLBA.}

\item{realized}{Logical scalar indicating whether to realize \code{object} into an
optimal memory layout for \code{IRLBA}. This improves computation time at the cost
of increased memory usage.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\code{\link[parallel:detectCores]{detectCores}}.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{name}{String specifying the name to be used to store the result in the
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} or
\code{\link[SeuratObject:Seurat-class]{reductions}} of the output.}

\item{layer}{Name of the layer to get from the assay data.}
}
\value{
\itemize{
\item \code{default} method: A numeric matrix where rows are cells and columns are
the two dimensions of the embedding.
\item \code{SingleCellExperiment} method: embedding was added into
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} named as \code{name}.
\item \code{Seurat} method: embedding was added into
\code{\link[SeuratObject:Seurat-class]{reductions}} named as \code{name}.
}
}
\description{
Perform Principal components analysis with \code{scrapper}
}
\seealso{
\code{\link[scrapper:runPca]{runPca}}
}
