% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runTSNE}
\alias{runTSNE}
\alias{runTSNE.default}
\alias{runTSNE.SingleCellExperiment}
\alias{runTSNE.Seurat}
\title{Compute the t-stochastic neighbor embedding with \code{scrapper}}
\usage{
runTSNE(object, ...)

\method{runTSNE}{default}(
  object,
  perplexity = 30L,
  n_neighbors = NULL,
  ...,
  max_depth = 20L,
  approximate = FALSE,
  max_iter = 500L,
  seed = 1234L,
  BNPARAM = AnnoyParam(),
  threads = NULL
)

\method{runTSNE}{SingleCellExperiment}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  name = "TSNE"
)

\method{runTSNE}{Seurat}(
  object,
  ...,
  dimred = "PCA",
  n_dimred = NULL,
  assay = NULL,
  layer = NULL,
  name = "TSNE"
)
}
\arguments{
\item{object}{A matrix-like object containing the counts (non-negative
integers). Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} methods.
}}

\item{perplexity}{Numeric scalar specifying the perplexity to use in the
t-SNE algorithm.}

\item{n_neighbors}{Integer scalar specifying the number of neighbors,
typically derived from \code{perplexity}.}

\item{max_depth}{Integer scalar specifying the maximum depth of the
Barnes-Hut quadtree. Smaller values (7-10) improve speed at the cost of
accuracy.}

\item{approximate}{Logical scalar indicating whether to use the
\dQuote{leaf approximation} approach, which sacrifices some accuracy for
greater speed. Only effective when \code{max_depth} is small enough for multiple
cells to be assigned to the same leaf node of the quadtree.}

\item{max_iter}{Integer scalar specifying the maximum number of iterations to
perform.}

\item{seed}{Integer scalar specifying the seed to use for generating the initial coordinates.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the
algorithm to use. Only used if \code{object} does not contain existing
nearest-neighbor results.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, all detected threads will be used. See
\code{\link[parallel:detectCores]{detectCores}}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{name}{String specifying the name to be used to store the result in the
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} or
\code{\link[SeuratObject:Seurat-class]{reductions}} of the output.}

\item{layer}{Name of the layer to get from the assay data.}
}
\value{
\itemize{
\item \code{default} method: A numeric matrix where rows are cells and columns are
the two dimensions of the embedding.
\item \code{SingleCellExperiment} method: embedding was added into
\code{\link[SingleCellExperiment:reducedDims]{reducedDims}} named as \code{name}.
\item \code{Seurat} method: embedding was added into
\code{\link[SeuratObject:Seurat-class]{reductions}} named as \code{name}.
}
}
\description{
Compute the t-stochastic neighbor embedding with \code{scrapper}
}
\seealso{
\code{\link[scrapper:runTsne]{runTsne}}
}
