% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFeatureSet.R
\name{scoreFeatureSet}
\alias{scoreFeatureSet}
\alias{scoreFeatureSet.default}
\alias{scoreFeatureSet.SummarizedExperiment}
\alias{scoreFeatureSet.Seurat}
\title{Score feature set activity for each cell}
\usage{
scoreFeatureSet(object, ...)

\method{scoreFeatureSet}{default}(
  object,
  feature_sets,
  ...,
  rank = 1,
  scale = FALSE,
  block = NULL,
  block_weight_policy = NULL,
  variable_block_weight = c(0, 1000),
  extra_work = 7,
  iterations = 1000,
  seed = NULL,
  realized = TRUE,
  threads = NULL
)

\method{scoreFeatureSet}{SummarizedExperiment}(object, ..., name = "scores", assay = "logcounts")

\method{scoreFeatureSet}{Seurat}(object, ..., assay = NULL, layer = "data")
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{\itemize{
\item \code{default} method: not used.
\item \code{SingleCellExperiment} and \code{Seurat} method: additional arguments passed to
\code{default} method.
}}

\item{feature_sets}{A list of integer, logical or character vector specifying
the features that belong to the set. You can also directly input an atomic
vector if there is only one feature set.}

\item{rank}{Integer scalar specifying the rank of the approximation.}

\item{scale}{Logical scalar indicating whether to scale rows to unit
variance.}

\item{block}{Factor specifying the block of origin (e.g., batch, sample) for
each cell in \code{x}. Alternatively \code{NULL} if all cells are from the same block.}

\item{block_weight_policy}{A string of \code{"variable"}, \code{"equal"}, or \code{"none"}
specifying the policy to use for weighting different blocks when computing
the average for each statistic Only used if \code{block} is not \code{NULL}.}

\item{variable_block_weight}{Numeric vector of length 2, specifying the
parameters for variable block weighting. The first and second values are used
as the lower and upper bounds, respectively, for the variable weight
calculation. Only used if \code{block} is not \code{NULL} and \code{block_weight_policy = "variable"}.}

\item{extra_work}{Integer scalar specifying the extra dimensions for the
IRLBA workspace.}

\item{iterations}{Integer scalar specifying the maximum number of restart iterations for IRLBA.}

\item{seed}{Integer scalar specifying the seed for the initial random vector in IRLBA.}

\item{realized}{Logical scalar indicating whether to realize \code{object} into an
optimal memory layout for \code{IRLBA}. This improves computation time at the cost
of increased memory usage.}

\item{threads}{Integer scalar specifying the number of threads to use. If
\code{NULL}, a single thread will be used.}

\item{name}{A string of the assay name for the scores.}

\item{assay}{Integer scalar or string indicating which assay of x
contains the expression values.}

\item{layer}{Name of the layer to get from the assay data.}
}
\value{
A matrix of feature scores.
}
\description{
Compute per-cell scores for a gene set, defined as the column sums of a rank-1 approximation to the submatrix for the gene set.
This uses the same approach implemented in the \pkg{GSDecon} package by Jason Hackney.
}
\seealso{
\code{\link[scrapper:scoreGeneSet]{scoreGeneSet}}
}
