% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subCluster.R
\name{subCluster}
\alias{subCluster}
\alias{subCluster.default}
\title{Find subclusters}
\usage{
subCluster(object, ...)

\method{subCluster}{default}(
  object,
  groups,
  restricted = NULL,
  ...,
  clusterFn = clusterSNNGraph,
  format = "\%s.\%s",
  new_level = "fill-end"
)
}
\arguments{
\item{object}{A matrix-like object. Rows are features and columns are cells.}

\item{...}{Additional arguments passed on to \code{clusterFn}.}

\item{groups}{A vector specifying the group assignment for each cell in
\code{object}.}

\item{restricted}{Character vector containing the subset of groups in
\code{groups} to be subclustered. By default, all unique groups in \code{groups} are
used for subclustering, but this can be restricted to specific groups of
interest to save compute time.}

\item{clusterFn}{A function returns a vector of cluster assignments for each
cell in that object.}

\item{format}{A string to be passed to \code{\link{sprintf}}, specifying how
the subclusters should be named with respect to the parent level in
\code{groups} and the level returned by \code{clusterFn}.}

\item{new_level}{A string of \code{"insert"}, \code{"fill-end"}, \code{"start"}, and \code{"end"}
indicates how to deal with the new levels. \code{"insert"} means the new levels
will be added into the old group level, \code{"fill-start"} and \code{"fill-end"} means
the first of the new levels will be insert into the old group level position
and the redundant new levels will be added in the start or end respectively.
\code{"start"} and \code{"end"} means the new levels will be added in the start or end,
and the old group level will be removed directly. Default: \code{"fill-end"}.}
}
\value{
An integer vector with cluster assignments for each cell. Additional
attributes return by \code{clusterFn} will be added for each subcluster run.
}
\description{
Find subclusters
}
