% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readH5AD.R
\name{readH5AD}
\alias{readH5AD}
\title{Read an H5AD File}
\usage{
readH5AD(file, delayed = TRUE)
}
\arguments{
\item{file}{A string specifying the path to the \code{.h5ad} file.}

\item{delayed}{A logical value indicating whether the assay data should be
loaded as \code{\link[HDF5Array:HDF5ArraySeed-class]{HDF5Array}} or
\code{\link[HDF5Array:H5SparseMatrix-class]{H5SparseMatrix}} objects from the \strong{HDF5Array}
package.}
}
\value{
A \link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}
object.
}
\description{
Read an H5AD File
}
